Add add&Remove.js and update index.js
