+++ index.js
@@ -4,64 +4,21 @@
 import './modules/addEvent.js';
 import './modules/time.js';
 import {
-  bookContainer,
-  bookTitle,
-  bookAuthor,
-  addBtn,
-} from './modules/variables.js';
-
-let bookList = [];
-
-class SingleBook {
-  constructor(title, author) {
-    this.title = title;
-    this.author = author;
-  }
-}
-class DisplayBook {
-  static bookUpload() {
-    bookContainer.innerHTML = '';
-    for (let i = 0; i < bookList.length; i += 1) {
-      bookContainer.innerHTML += `
-      <div class="books">
-        <div class ="book-wrapper"> 
-          <div class="book-holder">
-              <h3>"${bookList[i].title}"</h3>
-              <h3>by</h3>
-              <h3>${bookList[i].author}</h3>
-          </div>   
-          <button id='removeBtn'onclick="DisplayBook.deleted(${i})">Remove</button>
-        </div>   
-          <hr />
-      </div>
-     `;
-      bookTitle.value = '';
-      bookAuthor.value = '';
-    }
-  }
-
-  static deleted(index) {
-    bookList.splice(index, 1);
-    DisplayBook.bookUpload();
-    localStorage.setItem('bookList', JSON.stringify(bookList));
-  }
-
-  static addBook() {
-    const books = new SingleBook(bookTitle.value, bookAuthor.value);
-    bookList.push(books);
-  }
-}
+  pushToLocal,
+  showBook,
+  addBook,
+  bookUpload,
+} from './modules/add&Remove.js';
 
 window.onload = () => {
-  if (localStorage.getItem('bookList')) {
-    bookList = JSON.parse(localStorage.getItem('bookList'));
-  }
-  DisplayBook.bookUpload();
+  showBook();
 };
+
+const addBtn = document.querySelector('.add-btn');
 
 addBtn.addEventListener('click', (e) => {
   e.preventDefault();
-  DisplayBook.addBook();
-  DisplayBook.bookUpload();
-  localStorage.setItem('bookList', JSON.stringify(bookList));
+  addBook();
+  bookUpload();
+  pushToLocal();
 });
